Exif: cut the sidecar out of the tag reader
common/exif.cc was two unrelated files sharing a translation unit. One reads and
writes what a photograph says about itself -- EXIF, IPTC, XMP tags -- and names
nothing above layer 1. The other serialises the *development* into an XMP
sidecar: history stack, mask shapes, module order, and so reaches eleven
dt_ioppr_* symbols, dt_develop_blend_params_t, dt_masks_form_group_t and
dt_imageio_dng_write_tiff_header -- develop/ and imageio/, layers 5 and 6.

They meet only in the XMP document, which is what made the cut possible:

  metadata/exif.cc        2225 lines  27 tag functions + the 8-function seam
  common/xmp_sidecar.cc   2683 lines  36 sidecar functions
  metadata/exif_internal.h            class Lock + 3 functions, private to the two

The tag half can therefore live in the module, and `metadata: 0 includes from a
higher layer' still holds. The sidecar stays in common/ until src/history exists,
where it belongs -- it describes the development, not the photograph. Its five
entry points keep their dt_exif_* names so this cut stays reviewable; renaming
them is that move's business.

`#include "darktable.h"' is gone from both halves. Its only real use was
darktable.exiv2_threadsafe inside class Lock, which now goes through the
dt_exiv2_threadsafe_mutex() accessor that common/global_mutexes.h already
published; every other `darktable.' in the file is an XMP key string. Layering
violations 198 -> 196: that, and develop/imageop.h, which nothing needed.

Beyond pure code motion there are five edits and no others. _exif_decode_xmp_data
and _exif_read_exif_tag gain a dt_exif_ prefix (they stop being static, and a
leading underscore at global scope is reserved); those two and dt_remove_exif_keys
lose `static'; class Lock takes the mutex through the accessor; and
_exif_get_exiv2_tag_type reads the tag list through dt_exif_get_exiv2_taglist()
instead of the file-static it can no longer see -- the accessor builds the list if
it is empty, and dt_init() builds it long before any sidecar work, so the two are
identical in practice.

The cut itself was done by assigning every byte of the original to exactly one
destination and asserting, before writing anything, that consecutive segments abut
exactly with no gap or overlap, that all 73 parsed functions are placed and none
twice, and that the parts' line counts sum to 4775. A forward token search across
this file destroyed 117 lines once; the failure mode is silent. doc/exif-split.md
records the three ways the call graph lied before it was believed -- function
pointers, function-like macros as edges, and a `:(' in a comment parsing as a
parameter list.
