+++ src/CMakeLists.txt
@@ -76,7 +76,7 @@ FILE(GLOB SOURCE_FILES
   "common/dtpthread.c"
   "pixel/eaw.c"
   "system/memory_arena.c"
-  "common/exif.cc"
+  "common/xmp_sidecar.cc"
   "common/film.c"
   "common/file_location.c"
   "common/folder_survey.c"
@@ -149,6 +149,7 @@ FILE(GLOB SOURCE_FILES
   "common/utility.c"
   "common/variables.c"
   "metadata/colorlabels.c"
+  "metadata/exif.cc"
   "metadata/gpx.c"
   "metadata/map_locations.c"
   "metadata/metadata.c"
