+++ src/common/image.c
@@ -78,7 +78,8 @@
 #include "develop/imageop.h"
 #include "common/collection.h"
 #include "system/dtpthread.h"
-#include "common/exif.h"
+#include "common/xmp_sidecar.h"
+#include "metadata/exif.h"
 #include "common/file_location.h"
 #include "common/grouping.h"
 #include "common/history.h"
