+++ src/control/jobs/control_jobs.c
@@ -84,7 +84,8 @@
 #include "common/history_actions.h"
 #include "control/control.h"
 #include "common/collection.h"
-#include "common/exif.h"
+#include "common/xmp_sidecar.h"
+#include "metadata/exif.h"
 #include "common/film.h"
 #include "metadata/gpx.h"
 #include "common/history.h"
