+++ src/imageio/format/exr.cc
@@ -47,7 +47,7 @@
 #include "system/openmp.h"
 #include "system/mem_alloc.h"
 #include "common/module_versioning.h"
-#include "common/exif.h"
+#include "common/xmp_sidecar.h"
 #include "imageio/imageio_core.h"
 #include "imageio/imageio_module.h"
 #include "common/conf.h"
