+++ src/imageio/format/jpeg.c
@@ -45,7 +45,7 @@
 #include "common/module_versioning.h"
 #include <glib/gstdio.h>
 #include "caches/pixelpipe_cache_alloc.h"
-#include "common/exif.h"
+#include "metadata/exif.h"
 #include "imageio/imageio_core.h"
 #include "imageio/imageio_module.h"
 #include "common/conf.h"
