+++ src/imageio/imageio_core.c
@@ -67,7 +67,8 @@
 #include "common/times.h"
 #include "common/paths.h"
 #include "caches/pixelpipe_cache_alloc.h"
-#include "common/exif.h"
+#include "common/xmp_sidecar.h"
+#include "metadata/exif.h"
 #include "caches/image_cache.h"
 #include "common/history.h"
 #include "common/image_extensions.h"
