+++ src/imageio/imageio_exr.cc
@@ -52,7 +52,7 @@
 
 #include "system/macros.h"
 #include "system/openmp.h"
-#include "common/exif.h"
+#include "metadata/exif.h"
 #include "imageio/imageio_exr.h"
 #include "develop/develop.h"
