+++ src/imageio/imageio_jpeg.c
@@ -37,7 +37,7 @@
 #include "config.h"
 #include "caches/pixelpipe_cache_alloc.h"
 #endif
-#include "common/exif.h"
+#include "metadata/exif.h"
 #include "imageio/imageio_core.h"
 #include "imageio/imageio_jpeg.h"
 #include "develop/imageop.h"         // for IOP_CS_RGB
