+++ src/imageio/imageio_png.c
@@ -38,7 +38,7 @@
 #include "system/mem_alloc.h"
 #include "caches/pixelpipe_cache_alloc.h"
 #include "imageio_png.h"
-#include "common/exif.h"
+#include "metadata/exif.h"
 #include "develop/develop.h"
 
 int read_header(const char *filename, dt_imageio_png_t *png)
